Move model to better place
